Add a modal to confirm deletion of interview
There is some problem, the state management of the interviews of a candidate,
We should implement a pattern where state is lifted up to the parent component,
and changes are propagated down to the child through props i.e.
JobListingManager.tsx should host the state, because it seems this is the
first client component in the hierarchy, the parent of this component is
a server component.
We can also consider using valtio, or other related state management libs.