Handle syncing of the candidates list after creation
The problem was that, now that JobListingManager is hosting allCandidates,
any changes done by ManageCandidates is lifted up to JobListingManager,
and after JobListingManager is updated, ManageCandidates is rendered
with the updated allCandidates. The problem was because JobListingManager
was passing down props.allCandidates which is a copy of what it receives
from a server component, instead of passing allCandidates, which is a
state that it keeps within itself.